Refactoring chunks Add save to final file when read all file in first step (skip to write temporary file)

+++ src/main/java/org/example/context/ApplicationContext.java
@@ -11,7 +11,7 @@ public interface ApplicationContext {
 
   void sendSignal(IOException exception);
 
-  default OutputStream getOutputStream(int fileId) throws IOException {
+  default OutputStream getOutputStream(long fileId) throws IOException {
     var file = getFileSystemContext().getTemporaryFile(fileId);
     return getStreamFactory().getOutputStream(file);
   }
